requirements.txt
Flask>=3.0.0
google-genai>=0.2.0
Pillow>=10.2.0
Werkzeug>=3.0.1
